// src/app/models/EmployeePersonalDetails.ts
export class EmployeePersonalDetails {
    employeePersonalDetailsId: number;
    firstName: string;
    lastName: string;
    fatherName: string;
    cnicNo: string;
    gender: boolean;
    age: number;
    personalContactNo: string;
    education: string;
    currentAddress: string;
    permanentAddress: string;
    picture: string;
    createdTimestamp: string;

}